+++ Computer_graphics/lab_02/geometry/line.h
@@ -17,7 +17,7 @@ line_t *line_create(point_t *start, point_t *end);
 void line_free(line_t **line);
 
 line_t *line_move(line_t *line, point_t *offset);
-line_t *line_rotate(line_t *line, point_t *rotate_center, int angel);
+line_t *line_rotate(line_t *line, point_t *rotate_center, double angel);
 line_t *line_scale(line_t *line, point_t *scale_center, point_t *scale);
 
 #endif // LINE_H
